Project/frontend/requirements.txt
streamlit
requests==2.31.0   # For making HTTP requests
python-dotenv==1.0.0  # To load .env files